Convert away from pointer before converting to T. Repairs code for T == double.
git-svn-id: http://root.cern.ch/svn/root/trunk@47184 27541ba8-7e3a-0410-8455-c3a389f83636

+++ include/cling/Interpreter/Value.h
@@ -150,7 +150,7 @@ namespace cling {
         return (T) getAs<long double>();
       */
     } else if (desugCanon->isPointerType() || desugCanon->isObjectType()) {
-      return (T) getAs<void*>();
+      return (T) (size_t) getAs<void*>();
     }
     assert("unsupported type in Value, cannot cast simplistically!" && 0);
     return T();
